Move FullScreenQuad into class
